cli arguments (#7)
Fixes #2 

```
 go run main.go -h
NAME:
   rekor-phren is a tool to update the BigQuery table and the bucket with the rekor entries - rekor-phren update -u <rekor url> -b <bucket name> -t <table name> -s <start> -e <end> -c <concurrency>

USAGE:
   rekor-phren is a tool to update the BigQuery table and the bucket with the rekor entries [global options] command [command options] [arguments...]

COMMANDS:
   update   update -u <rekor url> -b <bucket name> -t <table name> -s <start> -e <end> -c <concurrency>
   help, h  Shows a list of commands or help for one command

GLOBAL OPTIONS:
   --bigquery-table-name value, -t value       (default: "rekor_test") [$PHREN_TABLE]
   --concurrency-level-update value, -c value  (default: 10) [$PHREN_CONCURRENCY]
   --end-index value, -e value                 (default: 0) [$PHREN_END]
   --gcs-bucket-name value, -b value           (default: openssf-rekor-test) [$PHREN_BUCKET]
   --help, -h                                  show help (default: false)
   --number-of-retries value, -r value         (default: 5) [$PHREN_RETRY]
   --rekor-url value, -u value                 (default: https://api.rekor.sigstore.dev) [$REKOR_URL]
   --start-index value, -s value               (default: 0) [$PHREN_START]

```

<172697+naveensrinivasan@users.noreply.github.com>

+++ pkg/Tlog.go
@@ -26,7 +26,7 @@ func NewTLog(host string) TLog {
 }
 
 // Size returns the size of the last entry.
-func (t *tlog) Size() (int64, error) {
+func (t *tlog) Size() (int, error) {
 	req, err := http.NewRequest("GET", fmt.Sprintf("%s/api/v1/log", t.host), nil)
 	if err != nil {
 		return 0, err
@@ -49,7 +49,7 @@ func (t *tlog) Size() (int64, error) {
 }
 
 // Entry returns the entry from the given tlogEntry.
-func (t *tlog) Entry(index int64) (Entry, error) {
+func (t *tlog) Entry(index int) (Entry, error) {
 	resp, err := http.Get(fmt.Sprintf("%s/api/v1/log/entries?logIndex=%d", t.host, index))
 	if err != nil {
 		return Entry{}, err
